requirements.txt
requests==2.26.0
Pillow==9.0.0
click==8.0.1
pandas==1.3.4
tqdm==4.62.3
pytorch-lightning==1.4.2
mlflow==1.23.1
scikit-learn==1.0.2
piqa==1.1.8